% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramashutility.R
\name{as.data.frame.stramash}
\alias{as.data.frame.stramash}
\title{Create data from from stramash object}
\usage{
\method{as.data.frame}{stramash}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{the fitted stramash object}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{not used, included for consistency as an S3 generic/method}

\item{...}{not used, included for consistency as an S3
generic/method.}
}
\description{
Creates data frame for easy plotting of results etc
}
\details{
Returns a data frame with named columnns
}
\examples{
a = stramash.workhorse(betahat = rnorm(100,0,2),sebetahat = rep(1, 100))
head(as.data.frame(a))

}

