% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramash.R
\name{autoselect.mixsd}
\alias{autoselect.mixsd}
\title{This is a copy from the ashr package because it is not exported and
I need it.}
\usage{
autoselect.mixsd(betahat, sebetahat, mult)
}
\arguments{
\item{betahat}{a p vector of estimates}

\item{sebetahat}{a p vector of corresponding standard errors}

\item{mult}{The amount to multiply by.}
}
\description{
This is a copy from the ashr package because it is not exported and
I need it.
}

