% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biash.R
\name{biash}
\alias{biash}
\title{Use control genes to estimate the unimodal inflation. Follow this
by running ash on the non-control genes.}
\usage{
biash(betahat, ctl, sebetahat, df = NULL, stramash_args = list())
}
\arguments{
\item{betahat}{a p vector of estimates}

\item{ctl}{A vector of logicals of length \code{length(betahat)}. A
\code{TRUE} at position \eqn{i} indicates that \eqn{\beta_i =
0}, i.e. that it is a control gene. A \code{FALSE} indicates
that it is not a control gene.}

\item{sebetahat}{a p vector of corresponding standard errors}

\item{df}{appropriate degrees of freedom for (t) distribution of
betahat/sebetahat if \code{likelihood = "t"}.}

\item{stramash_args}{A list containing additional arguments to pass to
\code{\link{stramash.workhorse}}}
}
\description{
Use control genes to estimate the unimodal inflation. Follow this
by running ash on the non-control genes.
}

