% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{calc_loglik_array}
\alias{calc_loglik_array}
\title{Calculate log-likelihood.}
\usage{
calc_loglik_array(g, betahat, errordist)
}
\arguments{
\item{g}{A mixture density. Either of class \code{normalmix} or of
class \code{unimix}. This is the prior.}

\item{betahat}{A vector of numerics. The locations at which to
evalutate the density.}

\item{errordist}{A list of objects of either class \code{normalmix}
or \code{unimix}. The length of this list must be the length of
\code{betahat}. \code{errordist[[i]]} is the \eqn{i}th error
distribution of \code{betahat[i]}.}
}
\description{
Calculate log-likelihood.
}
\author{
David Gerard
}

