% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{calc_nulllik_array}
\alias{calc_nulllik_array}
\title{Calculate the log-likelihood when all null.}
\usage{
calc_nulllik_array(betahat, errordist)
}
\arguments{
\item{betahat}{A vector of numerics. The locations at which to
evalutate the density.}

\item{errordist}{A list of objects of either class \code{normalmix}
or \code{unimix}. The length of this list must be the length of
\code{betahat}. \code{errordist[[i]]} is the \eqn{i}th error
distribution of \code{betahat[i]}.}
}
\description{
Calculate the log-likelihood when all null.
}
\author{
David Gerard
}

