% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramashutility.R
\name{cdf.stramash}
\alias{cdf.stramash}
\title{cdf method for stramash object}
\usage{
cdf.stramash(a, x, lower.tail = TRUE)
}
\arguments{
\item{a}{the fitted stramash object}

\item{x}{the vector of locations at which cdf is to be computed}

\item{lower.tail}{(default=TRUE) whether to compute the lower or upper tail}
}
\description{
Computed the cdf of the underlying fitted distribution
}
\details{
None
}

