% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{dnormalmix}
\alias{dnormalmix}
\title{Density function for a mixture of normals.}
\usage{
dnormalmix(x, mixdense)
}
\arguments{
\item{x}{The location to calculate the density.}

\item{mixdense}{An object of class \code{normalmix}.}
}
\description{
Density function for a mixture of normals.
}
\author{
David Gerard
}

