% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{dunimix}
\alias{dunimix}
\title{Density function for mixture of uniforms.}
\usage{
dunimix(x, mixdense)
}
\arguments{
\item{x}{The location to calculate the density.}

\item{mixdense}{An object of class \code{unimix}.}
}
\description{
Density function for mixture of uniforms.
}
\author{
David Gerard
}

