% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramash.R
\name{estimate_mixprop}
\alias{estimate_mixprop}
\title{Estimate mixture proportions of sigmaa by EM algorithm}
\usage{
estimate_mixprop(betahat, g, prior, optmethod = c("mixEM", "mixVBEM",
  "mixIP"), null.comp = 1, df = NULL, control = list(),
  errordist = NULL)
}
\arguments{
\item{betahat}{a p vector of estimates}

\item{g}{the prior distribution for beta (usually estimated from
the data; this is used primarily in simulated data to do
computations with the "true" g)}

\item{prior}{string, or numeric vector indicating Dirichlet prior
on mixture proportions (defaults to \code{"uniform"}, or
\eqn{(1,1...,1)}; also can be \code{"nullbiased"}
\eqn{(nullweight,1,...,1)} to put more weight on first
component), or \code{"unit"} \eqn{(1/K,...,1/K)} [for
\code{optmethod = mixVBEM} version only]}

\item{optmethod}{specifies optimization method used. Default is
\code{"mixIP"}, an interior point method, if REBayes is
installed; otherwise an EM algorithm is used. The interior
point method is faster for large problems (n>2000).}

\item{null.comp}{the position of the null component}

\item{df}{appropriate degrees of freedom for (t) distribution of
betahat/sebetahat if \code{likelihood = "t"}.}

\item{control}{A list of control parameters for the optmization
algorithm. Default value is set to be \code{control.default =
list(K = 1, method = 3, square = TRUE, step.min0 = 1, step.max0
= 1, mstep = 4, kr = 1, objfn.inc = 1, tol = 1.e-07, maxiter =
5000, trace = FALSE)}. User may supply changes to this list of
parameter, say, \code{control = list(maxiter = 10000, trace =
TRUE)}.}

\item{errordist}{A list of objects of either class
\code{\link[ashr]{normalmix}} or
\code{\link[ashr]{unimix}}. The length of this list must be the
length of \code{betahat}. Note that \code{errordist[[i]]} is
the \eqn{i}th error distribution of \code{betahat[i]}. Defaults
to \code{NULL}, in which case \code{stramash.workhorse} will
assume either a normal, t, or Laplace likelihood, depending on
the value for \code{likelihood}.}
}
\value{
A list, including the final loglikelihood, the null
    loglikelihood, a n by k likelihoodmatrix with (j,k)th element
    equal to \eqn{f_k(x_j)},and a flag to indicate convergence.
}
\description{
Estimate mixture proportions of sigmaa by EM algorithm
}

