% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramash.R
\name{gradient}
\alias{gradient}
\title{Copied from ashr package because I need it and it is not exported.}
\usage{
gradient(matrix_lik)
}
\arguments{
\item{matrix_lik}{The \eqn{(i, j)}th element of \code{matrix_like}
is the contribution to the likelihood of the \eqn{i}th
observation and the \eqn{j}th component.}
}
\value{
The kth element of this vector is the derivative
    of the loglik for \eqn{\pi=(\pi_0,...,1-\pi_0,...)} with respect to
    \eqn{\pi_0} at \eqn{\pi_0 = 1}.
}
\description{
Copied from ashr package because I need it and it is not exported.
}

