% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramash.R
\name{initpi}
\alias{initpi}
\title{This is copied from the ashr package because I need it and it is
not exported.}
\usage{
initpi(k, n, null.comp, randomstart)
}
\arguments{
\item{k}{The size of the grid.}

\item{n}{The sample size.}

\item{null.comp}{The position of the null component.}

\item{randomstart}{A logical. Should we start at a random location?}
}
\description{
This is copied from the ashr package because I need it and it is
not exported.
}

