% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{log_compdens_conv_mix}
\alias{log_compdens_conv_mix}
\title{Find log-component density convolution when error distribution is a
mixture.}
\usage{
log_compdens_conv_mix(g, betahat, errordist)
}
\arguments{
\item{g}{A mixture density. Either of class \code{normalmix} or of
class \code{unimix}. This is the prior.}

\item{betahat}{A vector of numerics. The locations at which to
evalutate the density.}

\item{errordist}{A list of objects of either class \code{normalmix}
or \code{unimix}. The length of this list must be the length of
\code{betahat}. \code{errordist[[i]]} is the \eqn{i}th error
distribution of \code{betahat[i]}.}
}
\value{
A matrix with row dimension \code{length(betahat)} and
    column dimension \code{length(errordist)}.
}
\description{
This function will find the log of the coefficients of the prior
mixing proportions in the ASH-likelihood when the model is
\eqn{betahat = g + errordist}, where \eqn{g} and \eqn{errordist}
are both mixtures of either normals or uniforms.
}
\details{
There are four cases: (1) \code{g} and \code{errordist} are both of
class \code{normalmix}, (2) \code{g} is of class \code{normalmix}
and \code{errordist} is of class \code{unifmix}, (3) \code{g} is of
class \code{unifmix} and \code{errordist} is of class
\code{normalmix}, and (4) \code{g} and \code{errordist} are both of
class \code{unifmix}. All of these are supported. Though (2) and
(3) differ only in indexing.
}
\author{
David Gerard
}

