% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{mix_cdf_array}
\alias{mix_cdf_array}
\title{Compute cdf from a mix_array.}
\usage{
mix_cdf_array(mixdist, q)
}
\arguments{
\item{mixdist}{Of class \code{unimix_array},
\code{normalmix_array}, or \code{truncnormalmix_array}. These
are the classes of the output from \code{\link{post_mix_dist}}.}

\item{q}{The quantile.}
}
\value{
The probability of being less than or equal to q.
}
\description{
Compute cdf from a mix_array.
}
\author{
David Gerard
}

