% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{mix_mean_array}
\alias{mix_mean_array}
\title{Compute mean from a mix_array.}
\usage{
mix_mean_array(mixdist)
}
\arguments{
\item{mixdist}{Of class \code{unimix_array},
\code{normalmix_array}, or \code{truncnormalmix_array}. These
are the classes of the output from \code{\link{post_mix_dist}}.}
}
\description{
Compute mean from a mix_array.
}
\author{
David Gerard
}

