% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{mix_probzero_array}
\alias{mix_probzero_array}
\title{Compute the probability of zero.}
\usage{
mix_probzero_array(mixdist)
}
\arguments{
\item{mixdist}{Of class \code{unimix_array},
\code{normalmix_array}, or \code{truncnormalmix_array}. These
are the classes of the output from \code{\link{post_mix_dist}}.}
}
\description{
Compute the probability of zero.
}
\author{
David Gerard
}

