% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{mix_sd_array}
\alias{mix_sd_array}
\title{Compute SD from a mix_array.}
\usage{
mix_sd_array(mixdist)
}
\arguments{
\item{mixdist}{Of class \code{unimix_array},
\code{normalmix_array}, or \code{truncnormalmix_array}. These
are the classes of the output from \code{\link{post_mix_dist}}.}
}
\value{
The posterior standard deviations of the location
    parameters.
}
\description{
Compute SD from a mix_array.
}
\author{
David Gerard
}

