% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{mixsd}
\alias{mixsd}
\title{Standard deviation of either class \code{\link[ashr]{normalmix}} or
    class \code{\link[ashr]{unimix}}}
\usage{
mixsd(m)
}
\arguments{
\item{m}{An object either of class \code{\link[ashr]{normalmix}} or
of class \code{\link[ashr]{unimix}}.}
}
\description{
Standard deviation of either class \code{\link[ashr]{normalmix}} or
    class \code{\link[ashr]{unimix}}
}
\author{
David Gerard
}

