% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{nnconv_dense}
\alias{nnconv_dense}
\title{Normal-normal convolution log-density evaluated at a single point.}
\usage{
nnconv_dense(g, errordist, x)
}
\arguments{
\item{g}{A \code{normalmix} object. This is the prior mixture.}

\item{errordist}{A \code{normalmix} object. This is the error mixture.}

\item{x}{A single numeric. This is the data.}
}
\description{
Normal-normal convolution log-density evaluated at a single point.
}
\author{
David Gerard
}
\seealso{
\code{\link{log_compdens_conv_mix.normalnormal}}
}

