% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramash.R
\name{normalize}
\alias{normalize}
\title{Copied from ashr package because I need it and it is not exported.}
\usage{
normalize(x)
}
\arguments{
\item{x}{A vector of numerics.}
}
\description{
Copied from ashr package because I need it and it is not exported.
}

