% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{nuconv_dense}
\alias{nuconv_dense}
\title{Normal-uniform convolution log-density evaluated at a single point.}
\usage{
nuconv_dense(g, errordist, x)
}
\arguments{
\item{g}{A \code{normalmix} object. This is the prior mixtrue.}

\item{errordist}{A \code{unimix} object. This is the error mixture.}

\item{x}{A single numeric. This is the data.}
}
\description{
Normal-uniform convolution log-density evaluated at a single point.
}
\author{
David Gerard
}
\seealso{
\code{\link{log_compdens_conv_mix.normaluni}}
}

