% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramash.R
\name{penloglik}
\alias{penloglik}
\title{Copied from ashr package because I need it and it is not exported.}
\usage{
penloglik(pi, matrix_lik, prior)
}
\arguments{
\item{pi}{A vector of porportions that sum to 1.}

\item{matrix_lik}{The \eqn{(i, j)}th element of \code{matrix_like}
is the contribution to the likelihood of the \eqn{i}th
observation and the \eqn{j}th component.}

\item{prior}{A vector of weights of the same length as \code{pi}.}
}
\description{
Copied from ashr package because I need it and it is not exported.
}

