% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{pnormalmix}
\alias{pnormalmix}
\title{CDF function for a mixture of normals.}
\usage{
pnormalmix(q, mixdense)
}
\arguments{
\item{q}{The quantile.}

\item{mixdense}{An object of class \code{normalmix}.}
}
\description{
CDF function for a mixture of normals.
}
\author{
David Gerard
}

