% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{post_mix_dist}
\alias{post_mix_dist}
\title{Generate the posterior distribution when the error distribution is
a mixture.}
\usage{
post_mix_dist(g, betahat, errordist)
}
\arguments{
\item{g}{The prior distribution. Either of class \code{"normalmix"}
or \code{"unimix"}.}

\item{betahat}{The observations, a vector of numerics.}

\item{errordist}{A list of mixture distributions.}
}
\value{
A list of arrays containing the parameters for the posterior.
The first dimension is the observations, the second is the mixture
components for the prior, and the third is the mixture components of
the error.

    \code{weights} The mixing proportions

    \code{means} The location parameter if the posterior is either
    a mixture of normals or truncated normals.

    \code{variances} The scale parameter if the posterior is either
    a mixture of normals or truncated normals.

    \code{lower} The lower bound of support if the posterior is
    either a mixture of truncated normals of uniforms.

    \code{upper} The upper bound of support if the posterior is
    either a mixture of truncated normals of uniforms.
}
\description{
Generate the posterior distribution when the error distribution is
a mixture.
}
\author{
David Gerard
}

