% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramashutility.R
\name{print.stramash}
\alias{print.stramash}
\title{Print method for stramash object}
\usage{
\method{print}{stramash}(x, ...)
}
\arguments{
\item{x}{the fitted stramash object}

\item{...}{not used, included for consistency as an S3
generic/method.}
}
\description{
Print the fitted distribution of beta values in the EB
    hierarchical model
}
\details{
None
}

