% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{punimix}
\alias{punimix}
\title{Probability function for mixture of uniforms.}
\usage{
punimix(p, mixdense)
}
\arguments{
\item{p}{The location to calculate the density.}

\item{mixdense}{An object of class \code{unimix}.}
}
\description{
Probability function for mixture of uniforms.
}
\author{
David Gerard
}

