% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{runimix}
\alias{runimix}
\title{Random draw from a mixture of uniforms.}
\usage{
runimix(n, mixdense)
}
\arguments{
\item{n}{the number of samples to draw.}

\item{mixdense}{An object of class \code{unimix}.}
}
\description{
Random draw from a mixture of uniforms.
}
\author{
David Gerard
}

