% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramash.R
\name{setprior}
\alias{setprior}
\title{This is a copy from the ashr package because it is not exported and
I need it.}
\usage{
setprior(prior, k, nullweight, null.comp)
}
\arguments{
\item{prior}{Either \code{"nullbiased"}, \code{"uniform"}, or
\code{"unit"}.}

\item{k}{The size of the grid.}

\item{nullweight}{The amount to penalize the null.}

\item{null.comp}{The position of the null component.}
}
\description{
This is a copy from the ashr package because it is not exported and
I need it.
}

