% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramash-package.r, R/stramash.R
\docType{package}
\name{stramash}
\alias{stramash}
\alias{stramash-package}
\title{stramash}
\usage{
stramash(betahat, errordist = NULL, sebetahat = NULL,
  mixcompdist = c("uniform", "halfuniform", "normal", "+uniform", "-uniform"),
  df = NULL, ...)
}
\arguments{
\item{betahat}{a p vector of estimates}

\item{errordist}{A list of objects of either class \code{normalmix}
or \code{unimix}. The length of this list must be the length of
\code{betahat}. \code{errordist[[i]]} is the \eqn{i}th error
distribution of \code{betahat[i]}. Defaults to \code{NULL}, in
which case \code{stramash.workhorse} will assume either a
normal or t likelihood, depending on the value for \code{df}.}

\item{sebetahat}{a p vector of corresponding standard errors}

\item{mixcompdist}{distribution of components in mixture
("uniform","halfuniform" or "normal"; "+uniform" or
"-uniform"), the default is "uniform". If you believe your
effects may be asymmetric, use "halfuniform". If you want to
allow only positive/negative effects use "+uniform"/"-uniform".
The use of "normal" is permitted only if df=NULL.}

\item{df}{appropriate degrees of freedom for (t) distribution of
betahat/sebetahat, default is NULL which is actually treated as
infinity (Gaussian)}

\item{...}{Further arguments to be passed to
\code{\link{stramash.workhorse}}.}
}
\value{
stramash returns an object of \code{\link[base]{class}} "stramash", a list with some or all of the following elements (determined by outputlevel) \cr
\item{fitted.g}{fitted mixture, either a normalmix or unimix}
\item{loglik}{log P(D|mle(pi))}
\item{logLR}{log[P(D|mle(pi))/P(D|beta==0)]}
\item{PosteriorMean}{A vector consisting the posterior mean of beta from the mixture}
\item{PosteriorSD}{A vector consisting the corresponding posterior standard deviation}
\item{PositiveProb}{A vector of posterior probability that beta is positive}
\item{NegativeProb}{A vector of posterior probability that beta is negative}
\item{ZeroProb}{A vector of posterior probability that beta is zero}
\item{lfsr}{The local false sign rate}
\item{lfdr}{A vector of estimated local false discovery rate}
\item{qvalue}{A vector of q values}
\item{call}{a call in which all of the specified arguments are specified by their full names}
\item{excludeindex}{the vector of index of observations with 0 standard error; if none, then returns NULL}
\item{model}{either "EE" or "ET", denoting whether exchangeable effects (EE) or exchangeable T stats (ET) has been used}
\item{optmethod}{the optimization method used}
\item{data}{a list consisting the input betahat and sebetahat (only included if outputlevel>2)}
}
\description{
The main function in the stramash package is
    \code{\link{stramash}}, which should be examined for more
    details. For simplicity only the most commonly-used options are
    documented under \code{\link{stramash}}. For expert or
    interested users the documentation for function
    \code{\link{stramash.workhorse}} provides documentation on all
    implemented options.

Takes vectors of estimates (betahat) and their
    standard errors (sebetahat), and applies shrinkage to them,
    using Empirical Bayes methods, to compute shrunk estimates for
    beta.
}
\details{
This function is actually just a simple wrapper that
    passes its parameters to \code{\link{stramash.workhorse}} which
    provides more documented options for advanced use. See readme
    for more details.
}
\examples{
beta = c(rep(0,100),rnorm(100))
sebetahat = abs(rnorm(200,0,1))
betahat = rnorm(200,beta,sebetahat)
beta.stramash = stramash.workhorse(betahat = betahat, sebetahat = sebetahat)
summary(beta.stramash)
names(beta.stramash)
graphics::plot(betahat,beta.stramash$PosteriorMean,xlim=c(-4,4),ylim=c(-4,4))

CIMatrix=stramashci(beta.stramash,level=0.95)
print(CIMatrix)

#Illustrating the non-zero mode feature
betahat=betahat+5
beta.stramash = stramash.workhorse(betahat = betahat, sebetahat = sebetahat)
graphics::plot(betahat,beta.stramash$PosteriorMean)
summary(beta.stramash)
}
\seealso{
\code{\link{stramash.workhorse}} for complete specification of
    stramash function

\code{\link{stramashci}} for computation of credible intervals
    after getting the stramash object return by \code{stramash()}

\code{\link{stramashm}} for Multi-model Adaptive Shrinkage
    function
}

