% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramashutility.R
\name{summary.stramash}
\alias{summary.stramash}
\title{Summary method for stramash object}
\usage{
\method{summary}{stramash}(object, ...)
}
\arguments{
\item{object}{the fitted stramash object}

\item{...}{not used, included for consistency as an S3
generic/method.}
}
\description{
Print summary of fitted stramash object
}
\details{
\code{\link{summary}} prints the fitted mixture, the
    fitted log likelihood with 10 digits and a flag to indicate
    convergence
}

