% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{t_to_mix}
\alias{t_to_mix}
\title{Generate an object of class \code{normalmix} that is an
approximation to a t-distribution.}
\usage{
t_to_mix(mu, sig, df, gridsize = 20)
}
\arguments{
\item{mu}{The mean of the t-distribution.}

\item{sig}{The scale parameter of the t-distribution.}

\item{df}{The degrees of freedom of the t-distribution.}

\item{gridsize}{The number of mixture components to use. The larger
the more accurate the approximation, but the higher the
computational load --- especially if you intend to use this for
\code{\link{ash.workhorse}}.}
}
\description{
This is based on the representation of the t as an inverse-gamma
scaled-mixture of Gaussians.
}
\author{
David Gerard
}

