% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stramash.R
\name{tic}
\alias{tic}
\title{From the ashr package because not exported but I need it.}
\usage{
tic(gcFirst = TRUE, type = c("elapsed", "user.self", "sys.self"))
}
\arguments{
\item{gcFirst}{A loogical.}

\item{type}{One of \code{"elapsed"}, \code{"user.self"}, or
\code{"sys.self"}.}
}
\description{
From the ashr package because not exported but I need it.
}

